requirements.txt
pygame==2.1.2
python-chess==1.999